requirements.txt
fastapi
uvicorn
streamlit
pandas
openpyxl
requests
python-dotenv
openai
pydantic
